Add a horizontal line below each consult--yank-read candidate
- The candidates do not include the horizontal line, it is only
displayed via text properties.

- The line comes at the end of the candidate, so I believe this should
not affect what selectrum displays at all. (This should be tested.)
So, this doesn't improve things for selectrum, but it should keep them
the same.

- Both icomplete-vertical and embark-occur display the full multiline
candidate. So consult + either can be regarded as a substitute for
browse-kill-ring. (We mentioned embark-occur already, I forgot that
icomplete-vertical shows multiline candidates too.)