Handle previews of invisible text
The seeming code duplication is necessary because previewing and jumping to a
position should be handled differently depending on user options
search-invisible, isearch-hide-immediately and consult-preview-mode.

save-excursion and save-restriction were moved in consult--preview-install so
that preview functions can assume original point position during cleanup.
This is necessary for isearch-clean-overlays to properly close every
temporarily opened overlay.