Make sure the grep commands are executed in the right directory
The grep should happen in the context of the directory returned by
`consult--directory-prompt`, i.e. the current project root if
`consult-project-root-function` is set.

This means two things, making sure `default-directory` is set correctly when the
command executes, and making sure `expand-file-name` is called with the right
directory so the relative paths are converted to absolute paths correctly.