fix #108: work around upstream bug
The upstream bug report is https://debbugs.gnu.org/cgi/bugreport.cgi?bug=45619.
The discussion there seems to go into the direction of disallowing dynamically
bound variables in patterns, and properly documenting that fact. I disagree with
those semantics of pcase and the broken optimization to remove seemingly unused
variables, in particular since it contradicts with the semantics of the
low-level let bindings. But we have to work with what we get :)