Fix broken match beginning/end when splitting OPTS
When `string-match` is used to determine whether we have OPTS or not, `(match-beginning 0)` and `(match-end 0)` are used to mark the split portion, get rid of the double dash and `substring` out the ARGS (i.e. `input`) and the OPTS (i.e. `opts`)

However, the subsequent call to `split-string-and-unquote` when setting `opts` contains an inner call to `string-match`, which clobbers the previous `(match-beginning 0)` value used to extract the `input`. For me, this resulted in the input being truncated to just 2 characters whenever there were OPTS provided. For example, when entering the input "foo -- -g *.el", the ARGS would just be "fo".

To fix this, just use a `let` to close over and remember the relevant match beginning and end values before they're overridden by `split-string-and-unquote`.