BREAKING CHANGE: consult-buffer - Enable preview by default for files/bookmarks
It is possible since 3f3663afa2b8ebff579d2239fd1b886dc06aa028 to configure the
:preview-key per virtual buffer source. This allows to set a manual preview key
for the slow buffer sources (files, bookmarks).

(dolist (src '(consult--source-file consult--source-project-file consult--source-bookmark))
  (set src (plist-put (symbol-value src) :preview-key (kbd "M-."))))

While this change is intrusive, Consult already activates preview everywhere
consistently. As described in the example configuration in the README, I
recommend to set a manual preview key for:

- consult-grep, consult-git-grep, consult-ripgrep
- consult-recent-file
- consult-bookmark
- consult--source-bookmark
- consult--source-file
- consult--source-project-file

The less-consistent alternatives are the following:

1. Avoid installing a preview function, where preview is potentially slow.
   This has been the case for the aforementioned consult--source-* sources.
   The downside is that then no manual preview can be configured.
2. Introduce another consult-preview-slow-key which is specificially used
   for the commands and sources, where preview is expected to be slow.