Transform input of consult-grep and consult-find commands
The input string is transformed before passing it to the backend. Spaces are
replaced by `.*`. Furthermore the input is interpreted as Emacs regular
expression and transformed to a Perl-compatible regular expression (PCRE), by a
handful of basic substitutions. The matching substrings are not obtained anymore
from the backend by looking for ANSI escape sequences. Instead the input regular
expression is used to heuristically apply highlighting.

Advantages:

- No parsing of ANSI escape sequences, which is fragile and inefficient.
- More flexibility, input string can be transformed, e.g., "def consult" -> "def.*consult".
- Highlighting becomes possible for backends which don't return highlights, e.g., find.

Disadvantages:

- Highlighting may be incorrect. In practice this is rare, since most search inputs
  are simple regular expressions.