consult-line: Introduce "cheap markers", upgraded on demand (Fix #517) (#519)
A cheap marker is a pair of a buffer and an integer position. We upgrade them
lazily on demand to real markers. Such an upgrade happens when changing the
window to perform recursive editing or when exporting the lines via
`embark-collect-*` or `embark-export`. `consult-line` generates a marker for
each line, which is very expensive even for moderately large files. The problem
is that the markers are updated until they are garbage collected. Furthermore
the markers are registered in a singly linked list (see marker.c), which leads
to performance issues when collecting many markers.