README: Prepare dropping consult-completing-read-multiple
Unfortunately it turned out that the consult-completing-read-multiple
implementation is fundamentally flawed (See #567). It is also impossible to save
the implementation since it was based on false assumptions. The assumption was
that the completion tables passed to completing-read-multiple don't have
non-trivial completion boundaries. For now I didn't remove the implementation
from Consult yet since there are a few packages like Citar, which still rely on
it. I will probably deprecate the command soon and remove it shortly after.