change(README): update Live Previews section for custom commands
If a custom command wraps a `consult` command, `this-command` is equal
to the name of the custom command, even when a `consult-*` command is
called within it. Since Consult uses the value of `this-command` to
determine whether or not to show a preview, we have to add the name of
our custom `consult-*` wrapping command to the `customize-consult`
call that sets the preview key.