Deprecate support for Selectrum
Selectrum doesn't fully support dynamic completion tables, which are widespread
in Emacs (find-file, completion-table-dynamic, ...) and has other technical
issues. Since there are multiple fully compliant completion UI alternatives
available, we phase out Selectrum support in Consult. Furthermore we consider
deprecating Selectrum in favor of Vertico. Vertico is simpler than Selectrum and
has been designed specifically to address the shortcomings of Selectrum. See
https://github.com/minad/vertico/issues/237 and
https://github.com/minad/vertico/wiki/Migrating-from-Selectrum-to-Vertico.

In the future the plan is to only support UIs which are fully compliant in
Consult. For the package ecosystem it is better to stick to fully compliant UIs,
since this reduces the amount of UI-specific workarounds across the ecosystem.

If you have feedback regarding the deprecation, please let me know. For the time
being, the deprecation message can be disabled by adding the following to your
init.el:

(with-eval-after-load 'consult-selectrum
  (advice-remove 'consult--read 'consult-selectrum--deprecated))