Deprecate consult-apropos
`consult-apropos' isn't a well-behaving command. Completion is used to select a
symbol, which is then passed to the `apropos' function. The `apropos' function
takes a regular expression as argument (but not a symbol). This mismatch leads
to two problems:

- The list of matches shown in the apropos buffer does not correspond to the
list of completion candidates.

- When entering a regular expression as input, the input is not passed to
`apropos' as long as the user does not explicitly submit the input via
`exit-minibuffer' or `vertico-exit-input'.

Therefore `consult-apropos' will be deprecated from now on. We have better tools
at our disposable via Embark. Try these command sequences:

- Export symbols to apropos: M-x describe-symbol <something> M-x embark-export
- Call apropos on selection: M-x describe-symbol <something> M-x embark-act a