Add user option `consult-kmacro-runner`.
This user option stores a function used to run the keyboard macro chosen in
`consult-kmacro`.  This can be useful, for example, with libraries that have the
ability to play back keyboard macros at multiple locations.

The function receives two arguments: the chosen keyboard macro (as a function)
and the prefix argument of `consult-kmacro`.