consult-register: Set register-command-info
Emacs recently changed how register jumping/loading is handled. The
variable register-use-preview may be used to adjust that behaviour;
it's read and acted upon by (e.g.) register-read-with-preview. That
function (which we call inside of consult-register-load) checks which
command called it and gets some information about that via the
register-command-info defmethod. If (say) the command indicates that
it just wants to jump to some register, an additional confirmation
is skipped when register-use-preview is set to an appropriate value.

Thus, to completely restore how Emacs used to handle register, we
need consult-register-load to signal that all it does it jump to
the register.