consult--async-split: do not insert initial split char if the minibuffer already starts with it
When running vertico-repeat-previous, vertico will clear the minibuffer and
then insert the text that was saved in history (e.g. "#some-text").

Later, when consult--async-split tries to inser the split char in the
minibuffer, the restored session has "##some-text" as input -- which does not
search for anything.

This commit fixes the issue by checking the minibuffer contents before
inserting the initial split char.